+++ adapter/__init__.py
@@ -5,20 +5,20 @@ Adapter module for image generation plugin
 
 from .gemini_adapter import GeminiAdapter
 from .gemini_openai_adapter import GeminiOpenAIAdapter
+from .gitee_ai_adapter import GiteeAIAdapter
 from .grok_adapter import GrokAdapter
 from .jimeng2api_adapter import Jimeng2APIAdapter
 from .openai_adapter import OpenAIAdapter
 from .siliconflow_adapter import SiliconFlowAdapter
 from .volcengine_ark_adapter import VolcengineArkAdapter
-from .z_image_adapter import ZImageAdapter
 
 __all__ = [
     "GeminiAdapter",
     "GeminiOpenAIAdapter",
+    "GiteeAIAdapter",
     "OpenAIAdapter",
     "SiliconFlowAdapter",
     "VolcengineArkAdapter",
-    "ZImageAdapter",
     "Jimeng2APIAdapter",
     "GrokAdapter",
 ]